DEV-92 move type definitions
